move settings components where they belong
