+++ frontend/src/metabase/admin/components/AdminContentTable/index.tsx
