+++ frontend/src/metabase/admin/components/AdminPaneLayout/AdminPaneLayout.tsx
