+++ frontend/src/metabase/admin/components/AdminPaneLayout/index.tsx
