+++ frontend/src/metabase/admin/people/containers/AdminPeopleApp/AdminPeopleApp.tsx
@@ -1,4 +1,4 @@
-import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
+import { AdminSettingsLayout } from "metabase/admin/components/AdminLayout/AdminSettingsLayout";
 
 import { PeopleNav } from "../../components/PeopleNav";
