+++ frontend/src/metabase/admin/performance/components/PerformanceApp.tsx
@@ -1,8 +1,8 @@
+import { AdminSettingsLayout } from "metabase/admin/components/AdminLayout/AdminSettingsLayout";
 import {
   AdminNavItem,
   AdminNavWrapper,
 } from "metabase/admin/components/AdminNav";
-import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
 import { PLUGIN_CACHING } from "metabase/plugins";
 
 import { PerformanceTabId } from "../types";
