+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/GroupMappingsWidgetView.tsx
@@ -4,8 +4,8 @@ import { useState } from "react";
 import { t } from "ttag";
 import _ from "underscore";
 
+import { AdminContentTable } from "metabase/admin/components/AdminContentTable";
 import { getErrorMessage } from "metabase/api/utils/errors";
-import { AdminContentTable } from "metabase/common/components/AdminContentTable";
 import CS from "metabase/css/core/index.css";
 import { FormSwitch } from "metabase/forms";
 import { Icon, Tooltip } from "metabase/ui";
