+++ frontend/src/metabase/admin/settingsRoutes.tsx
@@ -1,6 +1,6 @@
 import { IndexRedirect, Route } from "react-router";
 
-import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
+import { AdminSettingsLayout } from "metabase/admin/components/AdminLayout/AdminSettingsLayout";
 import { NotFound } from "metabase/common/components/ErrorPages";
 import {
   PLUGIN_AUTH_PROVIDERS,
